# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Node Comics API",
    "description": "私有漫画图片、持久化翻译任务、普通与 PLUS 会员权益及周期页数额度。",
    "version": "0.3.0"
  },
  "paths": {
    "/v1/me/queues": {
      "get": {
        "tags": [
          "queues"
        ],
        "summary": "Queues",
        "operationId": "queues_v1_me_queues_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/admin/users/{user_id}/queues": {
      "get": {
        "tags": [
          "queues"
        ],
        "summary": "User Queues",
        "operationId": "user_queues_v1_admin_users__user_id__queues_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/me/queues/{mode}/items": {
      "get": {
        "tags": [
          "queues"
        ],
        "summary": "Queue Items",
        "operationId": "queue_items_v1_me_queues__mode__items_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "mode",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "classic",
                "redraw"
              ],
              "type": "string",
              "title": "Mode"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/me/queues/{mode}/priority": {
      "post": {
        "tags": [
          "queues"
        ],
        "summary": "Priority",
        "operationId": "priority_v1_me_queues__mode__priority_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "mode",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "classic",
                "redraw"
              ],
              "type": "string",
              "title": "Mode"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PriorityRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/me/queues/{mode}/pause": {
      "post": {
        "tags": [
          "queues"
        ],
        "summary": "Pause",
        "operationId": "pause_v1_me_queues__mode__pause_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "mode",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "classic",
                "redraw"
              ],
              "type": "string",
              "title": "Mode"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PauseRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/me/translation-changes": {
      "get": {
        "tags": [
          "queues"
        ],
        "summary": "Changes",
        "operationId": "changes_v1_me_translation_changes_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PauseRequest": {
        "properties": {
          "paused": {
            "type": "boolean",
            "title": "Paused"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "paused"
        ],
        "title": "PauseRequest"
      },
...
      "PriorityRequest": {
        "properties": {
          "session_id": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Session Id"
          },
          "sequence": {
            "type": "integer",
            "maximum": 2147483647.0,
            "minimum": 0.0,
            "title": "Sequence"
          },
          "expected_version": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Expected Version"
          },
          "realtime_job_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Realtime Job Ids"
          },
          "ordered_job_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 500,
            "title": "Ordered Job Ids"
          },
          "ttl_seconds": {
            "type": "integer",
            "maximum": 300.0,
            "minimum": 1.0,
            "title": "Ttl Seconds",
            "default": 90
          },
          "takeover": {
            "type": "boolean",
            "title": "Takeover",
            "default": false
          },
          "reserve_next_upload": {
            "type": "boolean",
            "title": "Reserve Next Upload",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "session_id",
          "sequence",
          "expected_version"
        ],
        "title": "PriorityRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}